Rename wep routes to web
